Restructure away from templates
After tinkering with a few configurations, decided it would be best to
move away from mixing up the templating and whatever front end solution
I end up using. So, removed the EJS dependency and did some restructuring.

The express.js app will function mostly as an api server. The method I
used to set routes on the express app will be the same, and all of the
individual component resources of the site will still live in a single
api directory. However, the individual pieces of the site will all be
react components. No templating.