fix issue with error message middleware
The res.render method in express requires a view engine to be set,
which was not the case with the current app implementation. So the
error middleware was itself causing an error. switch it over to
just return json with the error message for now; may want to revisit
to make it a little friendlier for end users.